# swagger.json
{"openapi":"3.0.0","paths":{"/api/signup":{"post":{"operationId":"AuthController_signup","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SignupDto"}}}},"responses":{"201":{"description":""}},"summary":"Register a new user","tags":["Auth"]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"201":{"description":""}},"summary":"Login with email and password","tags":["Auth"]}},"/api/auth/me":{"get":{"operationId":"AuthController_getMe","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get current user profile","tags":["Auth"]}},"/api/rides":{"post":{"operationId":"RidesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateRideDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Create a ride request","tags":["Rides"]},"get":{"operationId":"RidesController_findAll","parameters":[{"name":"status","required":false,"in":"query","schema":{"type":"string"}},{"name":"driverId","required":false,"in":"query","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"List rides (role-based)","tags":["Rides"]}},"/api/rides/{id}":{"get":{"operationId":"RidesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get ride details","tags":["Rides"]}},"/api/rides/{id}/accept":{"patch":{"operationId":"RidesController_accept","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Driver accepts a ride","tags":["Rides"]}},"/api/rides/{id}/complete":{"patch":{"operationId":"RidesController_complete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Driver completes a ride","tags":["Rides"]}},"/api/rides/{id}/cancel":{"patch":{"operationId":"RidesController_cancel","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Tourist cancels a ride","tags":["Rides"]}},"/api/excursions":{"get":{"operationId":"ExcursionsController_findAll","parameters":[{"name":"category","required":false,"in":"query","schema":{"type":"string"}},{"name":"search","required":false,"in":"query","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}},{"name":"isActive","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"List excursions","tags":["Excursions"]},"post":{"operationId":"ExcursionsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateExcursionDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Create excursion (admin)","tags":["Excursions"]}},"/api/excursions/{id}":{"get":{"operationId":"ExcursionsController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get excursion details","tags":["Excursions"]},"patch":{"operationId":"ExcursionsController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateExcursionDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Update excursion (admin)","tags":["Excursions"]},"delete":{"operationId":"ExcursionsController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Soft-delete excursion (admin)","tags":["Excursions"]}},"/api/bookings":{"post":{"operationId":"BookingsController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateBookingDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Create a booking (tourist)","tags":["Bookings"]}},"/api/bookings/my":{"get":{"operationId":"BookingsController_findMy","parameters":[{"name":"status","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get my bookings (tourist)","tags":["Bookings"]}},"/api/bookings/pending":{"get":{"operationId":"BookingsController_findPending","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get pending bookings (guide)","tags":["Bookings"]}},"/api/bookings/guide-history":{"get":{"operationId":"BookingsController_findGuideHistory","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get guide booking history","tags":["Bookings"]}},"/api/bookings/all":{"get":{"operationId":"BookingsController_findAll","parameters":[{"name":"status","required":false,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get all bookings (admin)","tags":["Bookings"]}},"/api/bookings/{id}/confirm":{"patch":{"operationId":"BookingsController_confirm","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Confirm a booking (guide)","tags":["Bookings"]}},"/api/bookings/{id}/cancel":{"patch":{"operationId":"BookingsController_cancel","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Cancel a booking (tourist)","tags":["Bookings"]}},"/api/places":{"get":{"operationId":"PlacesController_findAll","parameters":[{"name":"type","required":false,"in":"query","schema":{"enum":["BEACH","RESTAURANT"],"type":"string"}},{"name":"search","required":false,"in":"query","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"List places (filterable by type)","tags":["Places"]},"post":{"operationId":"PlacesController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePlaceDto"}}}},"responses":{"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Create place (admin)","tags":["Places"]}},"/api/places/{id}":{"get":{"operationId":"PlacesController_findOne","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get place details","tags":["Places"]},"patch":{"operationId":"PlacesController_update","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePlaceDto"}}}},"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Update place (admin)","tags":["Places"]},"delete":{"operationId":"PlacesController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Soft-delete place (admin)","tags":["Places"]}},"/api/admin/stats":{"get":{"operationId":"AdminController_getStats","parameters":[],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Get dashboard stats","tags":["Admin"]}},"/api/admin/users":{"get":{"operationId":"AdminController_getUsers","parameters":[{"name":"userType","required":false,"in":"query","schema":{"type":"string"}},{"name":"search","required":false,"in":"query","schema":{"type":"string"}},{"name":"page","required":false,"in":"query","schema":{"type":"number"}},{"name":"limit","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"List all users","tags":["Admin"]}},"/api/admin/users/{id}/toggle":{"patch":{"operationId":"AdminController_toggleUser","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"security":[{"bearer":[]}],"summary":"Toggle user active status","tags":["Admin"]}}},"info":{"title":"MeiVolta API","description":"Tourism & Mobility super app API for Boa Vista island, Cape Verde","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"SignupDto":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com"},"password":{"type":"string","example":"password123","minLength":6},"name":{"type":"string","example":"John Doe"},"phone":{"type":"string","example":"+238 999 0000"},"userType":{"type":"string","enum":["TOURIST","DRIVER","GUIDE"],"example":"TOURIST"}},"required":["email","password","name","userType"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","example":"user@example.com"},"password":{"type":"string","example":"password123"}},"required":["email","password"]},"CreateRideDto":{"type":"object","properties":{"pickupLocation":{"type":"string","example":"Hotel Riu Karamboa"},"destination":{"type":"string","example":"Santa Monica Beach"},"passengers":{"type":"number","example":2,"minimum":1,"maximum":8},"serviceType":{"type":"string","enum":["TAXI","TRANSFER"],"example":"TAXI"},"scheduledTime":{"type":"string","example":"2025-06-15T10:00:00.000Z"},"notes":{"type":"string","example":"Please bring a child seat"}},"required":["pickupLocation","destination","passengers","serviceType"]},"CreateExcursionDto":{"type":"object","properties":{"nameEn":{"type":"string","example":"Beach Hopping Tour"},"namePt":{"type":"string","example":"Tour das Praias"},"descriptionEn":{"type":"string","example":"Discover the most beautiful beaches..."},"descriptionPt":{"type":"string","example":"Descubra as praias mais bonitas..."},"price":{"type":"number","example":4500},"duration":{"type":"string","example":"5 hours"},"category":{"type":"string","enum":["PICKUP_4X4","BEACH_TOUR","CULTURAL","TURTLES","BOAT","DIVING","FISHING","DESERT_VIANA","PHOTO_TOUR"]},"maxCapacity":{"type":"number","example":10},"imageUrl":{"type":"string","example":"https://www.jotform.com/blog/wp-content/uploads/2015/10/pexels-photo.jpg"}},"required":["nameEn","namePt","descriptionEn","descriptionPt","price","duration","category","maxCapacity"]},"UpdateExcursionDto":{"type":"object","properties":{"nameEn":{"type":"string"},"namePt":{"type":"string"},"descriptionEn":{"type":"string"},"descriptionPt":{"type":"string"},"price":{"type":"number"},"duration":{"type":"string"},"category":{"type":"string"},"maxCapacity":{"type":"number"},"imageUrl":{"type":"string"},"isActive":{"type":"boolean"}}},"CreateBookingDto":{"type":"object","properties":{"excursionId":{"type":"string","example":"uuid-here"},"date":{"type":"string","example":"2025-07-20T00:00:00.000Z"},"numberOfPeople":{"type":"number","example":2,"minimum":1}},"required":["excursionId","date","numberOfPeople"]},"CreatePlaceDto":{"type":"object","properties":{"nameEn":{"type":"string","example":"Santa Monica Beach"},"namePt":{"type":"string","example":"Praia de Santa Monica"},"descriptionEn":{"type":"string","example":"The longest beach..."},"descriptionPt":{"type":"string","example":"A praia mais longa..."},"type":{"type":"string","enum":["BEACH","RESTAURANT"]},"imageUrl":{"type":"string","example":"https://thehawaiivacationguide.com/wp-content/uploads/2021/10/Mala-Ocean-Tavern-where-to-eat-Lahaina-1024x576.jpg"},"location":{"type":"string","example":"South coast, 25min from Sal Rei"},"detailsEn":{"type":"object","example":{"cuisine":"Cape Verdean","hours":"12:00-22:00"}},"detailsPt":{"type":"object","example":{"cuisine":"Cabo-verdiana","hours":"12:00-22:00"}}},"required":["nameEn","namePt","descriptionEn","descriptionPt","type"]},"UpdatePlaceDto":{"type":"object","properties":{"nameEn":{"type":"string"},"namePt":{"type":"string"},"descriptionEn":{"type":"string"},"descriptionPt":{"type":"string"},"imageUrl":{"type":"string"},"location":{"type":"string"},"detailsEn":{"type":"object"},"detailsPt":{"type":"object"},"isActive":{"type":"boolean"}}}}}}